Move helper code into its own file, maybe
